% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghibli-package.R
\docType{package}
\name{ghibli-package}
\alias{ghibli}
\alias{ghibli-package}
\title{ghibli: Studio Ghibli Colour Palettes}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Colour palettes inspired by Studio Ghibli films, ported to R for your enjoyment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ewenme/ghibli}
  \item \url{https://ewenme.github.io/ghibli/}
  \item Report bugs at \url{https://github.com/ewenme/ghibli/issues}
}

}
\author{
\strong{Maintainer}: Ewen Henderson \email{ewenhenderson@gmail.com} (0000-0002-4748-4693)

Other contributors:
\itemize{
  \item Danielle Desrosiers [contributor]
}

}
\keyword{internal}
