% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghibli-package.R
\docType{package}
\name{ghibli}
\alias{ghibli}
\alias{ghibli-package}
\title{\code{ghibli} package}
\description{
Studio Ghibli Colour Palettes
}
\details{
See the README on
href{https://github.com/ewenme/ghibli}{GitHub}
}
