% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ghibli_palettes}
\alias{ghibli_palettes}
\title{Complete list of available ghibli palettes}
\format{A list of 27 character vector elements.}
\usage{
ghibli_palettes
}
\description{
Use \code{\link{ghibli_palette}} to construct palettes of desired length.
}
\keyword{datasets}
