% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ghibli_palettes}
\alias{ghibli_palettes}
\title{Complete list of available ghibli palettes}
\format{A list of 27 \code{color} objects elements.}
\usage{
ghibli_palettes
}
\description{
A list of all colour palettes to choose from, and preview directly in the console.
}
\details{
Pretty print and plot methods are powered by the \href{http://www.r-project.org}{prismatic}
package. Each palette is stored as a \code{color} object to enact this behaviour.
}
\seealso{
See \code{\link{ghibli_palette}} to construct palettes of desired length, and
\code{\link[prismatic]{colour}} for constructing color objects.
}
\keyword{datasets}
