% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\docType{data}
\name{ghibli_palettes}
\alias{ghibli_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 21.}
\usage{
ghibli_palettes
}
\description{
Use \code{\link{ghibli_palette}} to construct palettes of desired length.
}
\keyword{datasets}
